Intial commit with the skeleton framwork
Script.js has all the basics of the framework at the moment, will eventually break those up into a more logical file structure instead of the current implementation. Index.html is the obvious basic html of the project and i added Angular.js file to eventually convert the frontend to angular but that may change slightly and that is a little ways away as i want to have the framework alittle more esablished before moving to the front end.